Battle refactor

+++ entities/boss/behavior/BossBehaviorStrategy.java
@@ -1,5 +1,5 @@
 package entities.boss.behavior;
 
 public interface BossBehaviorStrategy {
-    double calculateDamage();
+    int calculateDamage();
 }
